npc system file restructure waiter idling outline
