+++ src/engine/components/interactable_component.cpp
@@ -1,41 +1,9 @@
 #include "component.h"
-#include "config.h"
 #include "debug_log_interface.h"
 #include "raglib.h"
 #include "system.h"
+#include <string>
 
-bool components::interactable_component::can_accept_interactor() const{
-    for(size_t i = 0; i < positions_.size(); ++i){
-        if(positions_[i].has_value() and not interactors_[i].has_value()){
-            return true;
-        }
-    }
-    return false;
-}
-bool components::interactable_component::claim(size_t interactor_id){
-    for(auto& interactor : interactors_){
-        if(interactor.has_value() and interactor.value() == interactor_id){
-            return false;
-        }
-    }
-    for(size_t i = 0; i < positions_.size(); ++i){
-        if(positions_[i].has_value() and not interactors_[i].has_value()){
-            interactors_[i] = interactor_id;
-            return true;
-        }
-    }
-    return false;
-}
-void components::interactable_component::release(size_t interactor_id){
-    for(auto& interactor : interactors_){
-        if(interactor.has_value() and interactor.value() == interactor_id){
-            interactor.reset();
-        }
-    }
-}
-const std::array<std::optional<size_t>, DIRECTIONS>& components::interactable_component::get_interactors() const{
-    return interactors_;
-}
 Rectangle components::interactable_component::get_interaction_box(Rectangle box) const{
     return Rectangle{box.x - reach_,
                      box.y - reach_,
@@ -84,4 +52,36 @@ std::optional<Vector2> components::interactable_component::get_interaction_offse
         ? "[interactable_component::get_interaction_offset] chosen offset: " + raglib::vector_to_string(closest_offset.value())
         : "[interactable_component::get_interaction_offset] no free interaction slot found");
     return closest_offset;
+}
+bool components::interactable_component::can_accept_interactor() const{
+    for(size_t i = 0; i < positions_.size(); ++i){
+        if(positions_[i].has_value() and not interactors_[i].has_value()){
+            return true;
+        }
+    }
+    return false;
+}
+bool components::interactable_component::claim(size_t interactor_id){
+    for(auto& interactor : interactors_){
+        if(interactor.has_value() and interactor.value() == interactor_id){
+            return false;
+        }
+    }
+    for(size_t i = 0; i < positions_.size(); ++i){
+        if(positions_[i].has_value() and not interactors_[i].has_value()){
+            interactors_[i] = interactor_id;
+            return true;
+        }
+    }
+    return false;
+}
+void components::interactable_component::release(size_t interactor_id){
+    for(auto& interactor : interactors_){
+        if(interactor.has_value() and interactor.value() == interactor_id){
+            interactor.reset();
+        }
+    }
+}
+const std::array<std::optional<size_t>, DIRECTIONS>& components::interactable_component::get_interactors() const{
+    return interactors_;
 }
