+++ src/engine/components/movement_component.cpp
@@ -1,5 +1,6 @@
 #include "component.h"
 #include "config.h"
+#include <raymath.h>
 
 bool components::movement_component::has_reached_position(Vector2 position){
     if(paths_.empty()){
@@ -10,15 +11,6 @@ bool components::movement_component::has_reached_position(Vector2 position){
 }
 path::path& components::movement_component::get_current_path(){
     return paths_.empty() ? path::empty_path : paths_.front();
-}
-Vector2 components::movement_component::get_move_speed(){
-    return move_speed_;
-}
-Vector2 components::movement_component::get_direction_scalar(){
-    return direction_scalar_;
-}
-void components::movement_component::set_direction_scalar(Vector2 direction_scalar){
-    direction_scalar_ = direction_scalar;
 }
 std::queue<path::path>& components::movement_component::get_paths(){
     return paths_;
@@ -43,4 +35,13 @@ void components::movement_component::clear_paths(){
 void components::movement_component::finish_path(){
     if(paths_.empty()){ return; }
     paths_.pop();
+}
+Vector2 components::movement_component::get_move_speed(){
+    return move_speed_;
+}
+Vector2 components::movement_component::get_direction_scalar(){
+    return direction_scalar_;
+}
+void components::movement_component::set_direction_scalar(Vector2 direction_scalar){
+    direction_scalar_ = direction_scalar;
 }
