+++ src/engine/components/storage_component.cpp
